Fix #24, fix #27, fix #28. Move tests, change package name, add multi-stage support
