+++ .idea/runConfigurations/Docker_Compose.xml
@@ -2,6 +2,7 @@
   <configuration default="false" name="Docker Compose" type="docker-deploy" factoryName="docker-compose.yml" server-name="Docker">
     <deployment type="docker-compose.yml">
       <settings>
+        <option name="buildKitEnabledForCompose" value="true" />
         <option name="envFilePath" value="" />
         <option name="sourceFilePath" value="compose.yaml" />
       </settings>
